+++ tsup.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   entry: [
     'src/library/maplibre-gl-three.ts',
     'src/library/proj4-wrapper.ts',
+    'src/library/geotiff-wrapper.ts',
   ],
   format: ['esm'],
   dts: true,
@@ -13,6 +14,7 @@ export default defineConfig({
   external: [
     '3d-tiles-renderer',
     'maplibre-gl',
+    'geotiff',
     'proj4',
     'three',
     /^three\//,