+++ www/examples/basic/maplibreGlThree-npm-example/src/script.ts
@@ -21,9 +21,9 @@ const map = new Map({
 
 map.addControl(new NavigationControl({ visualizePitch: true }));
 
-map.on('load', () => {
+map.on('load', async () => {
     const threeDManager = new ThreeDManager();
-    const agiHqTiles = threeDManager.load3dTiles({
+    const agiHqTiles = await threeDManager.load3dTiles({
         tilesetUrl: 'https://pelican-public.s3.amazonaws.com/3dtiles/agi-hq/tileset.json',
         layerId: 'agi-hq-3d-tiles',
         offset: { east: 0, up: -310, south: 0 },