+++ www/examples/basic/maplibreGlThree-selfhost-example/script.js
@@ -13,15 +13,15 @@ const map = new maplibregl.Map({
 
 map.addControl(new maplibregl.NavigationControl({ visualizePitch: true }));
 
-map.on('load', () => {
+map.on('load', async () => {
     const threeDManager = new ThreeDManager({
         dracoPath: "/dependencies/three@0.183.2/examples/jsm/libs/draco/",
         ktx2Path: "/dependencies/three@0.183.2/examples/jsm/libs/basis/"
     });
-    const agiHqTiles = threeDManager.load3dTiles({
+    const agiHqTiles = await threeDManager.load3dTiles({
         tilesetUrl: '/datasets/agi-hq/tileset.json',
         layerId: 'agi-hq-3d-tiles',
-        offset: { east: 0, up: -310, south: 0 },
+        offset: { east: 0, up: -234, south: 0 },
     });
 
     map.addLayer(agiHqTiles.getLayer(), "rivers");